I upgraded react-native to 0.48.1 and fixed the issues for both Android and iOS.
After the change, I could run the simulator with the application for both Android and iOS.

upgrading package versions made it work better with the latest Android SDK
Tools like emulator and adb are used in the background by react-native. So upgrading is good.

I fixed the dependency problem for react-test-renderer
I fixed the PropTypes warning when runnning the tests

iOS:
change of one file path
removal of parallelized builds
fix the need for an optional dependency